move theme to shared ui module
