--- samples/repo-search/desktop-app/src/jvmMain/kotlin/RepoSearchTheme.kt
@@ -1,16 +1,0 @@
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.darkColors
-import androidx.compose.material.lightColors
-import androidx.compose.runtime.Composable
-
-@Composable
-fun RepoSearchTheme(
-  darkTheme: Boolean = isSystemInDarkTheme(),
-  content: @Composable () -> Unit,
-) {
-  MaterialTheme(
-    colors = if (darkTheme) darkColors() else lightColors(),
-    content = content,
-  )
-}
