+++ samples/repo-search/shared/build.gradle.kts
@@ -1,5 +1,8 @@
+@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
 plugins {
   alias(libs.plugins.kotlin.multiplatform)
+  alias(libs.plugins.jetbrains.compose)
+  id("com.android.library")
   alias(libs.plugins.kotlin.serialization)
   alias(libs.plugins.kotlin.native.cocoapods)
 }
@@ -8,6 +11,7 @@ kotlin {
   ios()
   iosSimulatorArm64()
   jvm()
+  android()
 
   cocoapods {
     summary = "Shared code for Repo Search."
@@ -22,6 +26,10 @@ kotlin {
         implementation(libs.ktor.client.core)
         implementation(libs.ktor.client.contentNegotiation)
         implementation(libs.ktor.serialization.kotlinx.json)
+        implementation(compose.ui)
+        implementation(compose.foundation)
+        implementation(compose.material)
+        implementation(compose.runtime)
       }
     }
     val iosMain by getting {
