+++ samples/repo-search/shared/gradle.properties
@@ -1,0 +1,1 @@
+org.jetbrains.compose.experimental.uikit.enabled=true
